// backend/src/modules/production-orders/dto/create-production-order-bottle-volume-check.dto.ts
export class CreateProductionOrderBottleVolumeCheckDto {
  bottle_1_volume?: number | string | null;
  bottle_2_volume?: number | string | null;
  bottle_3_volume?: number | string | null;
  bottle_4_volume?: number | string | null;
  bottle_5_volume?: number | string | null;
  bottle_6_volume?: number | string | null;
}